thinking about graph hash:
- not sure what the best default for dealing with supercells should be
- should mofcecker just make niggli cells for everything?
